requirements.txt
djangorestframework == 3.15.2
soundfile
transformers
pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
qrcode[pil]